Fixed injection point handler
Identifying injection points make it possible to encode the payload differently depending on its injection point position. This change adds injection point identification logic so that url encoding is applied on the request line and only optionally in other locations. This change make more granular tests possible. One future example could be to attack HTTP::uri by not starting the request path parameter with a / but instead with the "attackpattern".
This change was contributed largely by Matthias Karlsson.